requirements.txt
fastapi
uvicorn[standard]
openai
pdfminer.six
python-docx
openpyxl
python-pptx
